requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
openai==1.30.1
httpx==0.27.0
python-dotenv==1.0.1